+++ packages/zenrouter_core/lib/src/coordinator/base.dart
@@ -436,7 +436,7 @@ abstract class CoordinatorCore<T extends RouteUri> extends Equatable
       }
     }
 
-    return false;
+    return null;
   }
 
   /// Triggers a rebuild of the coordinator.